First model for events

+++ members/services.py
@@ -598,8 +598,9 @@ def person_resign(person):
 
 
 def person_deregister(person):
-    user = person.auth
-    user.delete()
+    if person.auth:
+        user = person.auth
+        user.delete()
     person.auth = None
     person.pin = None
     person.allow_phone = False
